Begin big refactor to bring structure to the code and make it expandable
