+++ maps/test.json
@@ -1,0 +1,15 @@
+{
+    "map": [
+        [0, 1, 0, 0, 0, 0, 0],
+        [0, 1, 0, 1, 1, 1, 0],
+        [0, 1, 1, 1, 0, 1, 0],
+        [0, 0, 0, 0, 0, 1, 0],
+        [1, 1, 1, 0, 1, 1, 0],
+        [0, 0, 1, 1, 1, 0, 0],
+        [0, 0, 0, 0, 0, 0, 0]
+    ],
+    "IDs": {
+        "0": "platform",
+        "1": "path"
+    }
+}
